Deprecated Functions

Archive of deprecated functions. Some of these might be removed from the package in later releases.

Deprecated methods for building the data sets needed for plotting roc and prc plots. use confusion_matrix instead.

Usage

qroc_build_data_frame(fit, truth = NULL, n_threshold = 200, ...)

# Default S3 method
qroc_build_data_frame(fit, truth = NULL, n_threshold = 200, ...)

# S3 method for class 'glm'
qroc_build_data_frame(fit, truth = NULL, n_threshold = 200, ...)

qprc_build_data_frame(fit, n_threshold = 200, ...)

Arguments

fit

a glm fit with family = binomial(), or predicted values

truth

ignored if fit is a glm object. A vector of observations, 0/1 or FALSE/TRUE values, of equal length to fit

n_threshold

number of thresholds to use to estimate auroc or auprc

...

passed to predict