Qable: an extended version of knitr::kable

Create a simple table via kable with row groups and rownames similar to those of latex from the Hmisc package or htmlTable from the htmlTable package.

Usage

qable(
  x,
  rtitle = "",
  rgroup = numeric(0),
  rnames = rownames(x),
  cnames = colnames(x),
  markup = getOption("qwraps2_markup", "latex"),
  kable_args = list(),
  ...
)

Arguments

x

matrix or data.frame to be turned into a qable

rtitle

a row grouping title. See Details.

rgroup

a named numeric vector with the name of the row group and the number of rows within the group. sum(rowgroup) == nrow(x).

rnames

a character vector of the row names

cnames

column names

markup

the markup language to use expected to be either "markdown" or "latex"

kable_args

a list of named arguments to send to kable. See Details.

...

pass through

Value

qable returns a qwraps2_qable object that is just a character matrix with some additional attributes and the print method returns, invisibly, the object passed to print.

Details

rtitle can be used to add a title to the column constructed by the rgroup and rnames. The basic layout of a table generated by qable is:

rtitle	cnames[1]	cnames[2]

rgroup[1]

rnames[1]	x[1, 1]	x[1, 2]

rnames[2]	x[2, 1]	x[2, 2]

rnames[3]	x[3, 1]	x[3, 2]

rgroup[2]

rnames[4]	x[4, 1]	x[4, 1]

rnames[5]	x[5, 1]	x[5, 1]

Passing arguments to link[knitr]{kable} is done via the list kable_args. This is an improvement in 0.6.0 to address arguments with different use between qable and kable but the same name, notably format. Within the print method for qwraps2_qable objects, some default arguments for knitr::kable are created.

Defaults if the named element of kable_args is missing: kable_args$format will be "latex" if markup = "latex" and will be "pipe" if markup = "markdown".

kable_args$escape = !(markup = "latex")

kable_args$row.names defaults to FALSE

kable_args$col.names defaults to colnames(x)

See also

Examples

data(mtcars)
x <- qable(mtcars)
x
#> 
#> \begin{tabular}{l|l|l|l|l|l|l|l|l|l|l|l}
#> \hline
#>  & mpg & cyl & disp & hp & drat & wt & qsec & vs & am & gear & carb\\
#> \hline
#> Mazda RX4 & 21 & 6 & 160 & 110 & 3.9 & 2.62 & 16.46 & 0 & 1 & 4 & 4\\
#> \hline
#> Mazda RX4 Wag & 21 & 6 & 160 & 110 & 3.9 & 2.875 & 17.02 & 0 & 1 & 4 & 4\\
#> \hline
#> Datsun 710 & 22.8 & 4 & 108 & 93 & 3.85 & 2.32 & 18.61 & 1 & 1 & 4 & 1\\
#> \hline
#> Hornet 4 Drive & 21.4 & 6 & 258 & 110 & 3.08 & 3.215 & 19.44 & 1 & 0 & 3 & 1\\
#> \hline
#> Hornet Sportabout & 18.7 & 8 & 360 & 175 & 3.15 & 3.44 & 17.02 & 0 & 0 & 3 & 2\\
#> \hline
#> Valiant & 18.1 & 6 & 225 & 105 & 2.76 & 3.46 & 20.22 & 1 & 0 & 3 & 1\\
#> \hline
#> Duster 360 & 14.3 & 8 & 360 & 245 & 3.21 & 3.57 & 15.84 & 0 & 0 & 3 & 4\\
#> \hline
#> Merc 240D & 24.4 & 4 & 146.7 & 62 & 3.69 & 3.19 & 20 & 1 & 0 & 4 & 2\\
#> \hline
#> Merc 230 & 22.8 & 4 & 140.8 & 95 & 3.92 & 3.15 & 22.9 & 1 & 0 & 4 & 2\\
#> \hline
#> Merc 280 & 19.2 & 6 & 167.6 & 123 & 3.92 & 3.44 & 18.3 & 1 & 0 & 4 & 4\\
#> \hline
#> Merc 280C & 17.8 & 6 & 167.6 & 123 & 3.92 & 3.44 & 18.9 & 1 & 0 & 4 & 4\\
#> \hline
#> Merc 450SE & 16.4 & 8 & 275.8 & 180 & 3.07 & 4.07 & 17.4 & 0 & 0 & 3 & 3\\
#> \hline
#> Merc 450SL & 17.3 & 8 & 275.8 & 180 & 3.07 & 3.73 & 17.6 & 0 & 0 & 3 & 3\\
#> \hline
#> Merc 450SLC & 15.2 & 8 & 275.8 & 180 & 3.07 & 3.78 & 18 & 0 & 0 & 3 & 3\\
#> \hline
#> Cadillac Fleetwood & 10.4 & 8 & 472 & 205 & 2.93 & 5.25 & 17.98 & 0 & 0 & 3 & 4\\
#> \hline
#> Lincoln Continental & 10.4 & 8 & 460 & 215 & 3 & 5.424 & 17.82 & 0 & 0 & 3 & 4\\
#> \hline
#> Chrysler Imperial & 14.7 & 8 & 440 & 230 & 3.23 & 5.345 & 17.42 & 0 & 0 & 3 & 4\\
#> \hline
#> Fiat 128 & 32.4 & 4 & 78.7 & 66 & 4.08 & 2.2 & 19.47 & 1 & 1 & 4 & 1\\
#> \hline
#> Honda Civic & 30.4 & 4 & 75.7 & 52 & 4.93 & 1.615 & 18.52 & 1 & 1 & 4 & 2\\
#> \hline
#> Toyota Corolla & 33.9 & 4 & 71.1 & 65 & 4.22 & 1.835 & 19.9 & 1 & 1 & 4 & 1\\
#> \hline
#> Toyota Corona & 21.5 & 4 & 120.1 & 97 & 3.7 & 2.465 & 20.01 & 1 & 0 & 3 & 1\\
#> \hline
#> Dodge Challenger & 15.5 & 8 & 318 & 150 & 2.76 & 3.52 & 16.87 & 0 & 0 & 3 & 2\\
#> \hline
#> AMC Javelin & 15.2 & 8 & 304 & 150 & 3.15 & 3.435 & 17.3 & 0 & 0 & 3 & 2\\
#> \hline
#> Camaro Z28 & 13.3 & 8 & 350 & 245 & 3.73 & 3.84 & 15.41 & 0 & 0 & 3 & 4\\
#> \hline
#> Pontiac Firebird & 19.2 & 8 & 400 & 175 & 3.08 & 3.845 & 17.05 & 0 & 0 & 3 & 2\\
#> \hline
#> Fiat X1-9 & 27.3 & 4 & 79 & 66 & 4.08 & 1.935 & 18.9 & 1 & 1 & 4 & 1\\
#> \hline
#> Porsche 914-2 & 26 & 4 & 120.3 & 91 & 4.43 & 2.14 & 16.7 & 0 & 1 & 5 & 2\\
#> \hline
#> Lotus Europa & 30.4 & 4 & 95.1 & 113 & 3.77 & 1.513 & 16.9 & 1 & 1 & 5 & 2\\
#> \hline
#> Ford Pantera L & 15.8 & 8 & 351 & 264 & 4.22 & 3.17 & 14.5 & 0 & 1 & 5 & 4\\
#> \hline
#> Ferrari Dino & 19.7 & 6 & 145 & 175 & 3.62 & 2.77 & 15.5 & 0 & 1 & 5 & 6\\
#> \hline
#> Maserati Bora & 15 & 8 & 301 & 335 & 3.54 & 3.57 & 14.6 & 0 & 1 & 5 & 8\\
#> \hline
#> Volvo 142E & 21.4 & 4 & 121 & 109 & 4.11 & 2.78 & 18.6 & 1 & 1 & 4 & 2\\
#> \hline
#> \end{tabular}
qable(mtcars, markup = "markdown")
#> 
#> 
#> |                    |mpg  |cyl |disp  |hp  |drat |wt    |qsec  |vs |am |gear |carb |
#> |:-------------------|:----|:---|:-----|:---|:----|:-----|:-----|:--|:--|:----|:----|
#> |Mazda RX4           |21   |6   |160   |110 |3.9  |2.62  |16.46 |0  |1  |4    |4    |
#> |Mazda RX4 Wag       |21   |6   |160   |110 |3.9  |2.875 |17.02 |0  |1  |4    |4    |
#> |Datsun 710          |22.8 |4   |108   |93  |3.85 |2.32  |18.61 |1  |1  |4    |1    |
#> |Hornet 4 Drive      |21.4 |6   |258   |110 |3.08 |3.215 |19.44 |1  |0  |3    |1    |
#> |Hornet Sportabout   |18.7 |8   |360   |175 |3.15 |3.44  |17.02 |0  |0  |3    |2    |
#> |Valiant             |18.1 |6   |225   |105 |2.76 |3.46  |20.22 |1  |0  |3    |1    |
#> |Duster 360          |14.3 |8   |360   |245 |3.21 |3.57  |15.84 |0  |0  |3    |4    |
#> |Merc 240D           |24.4 |4   |146.7 |62  |3.69 |3.19  |20    |1  |0  |4    |2    |
#> |Merc 230            |22.8 |4   |140.8 |95  |3.92 |3.15  |22.9  |1  |0  |4    |2    |
#> |Merc 280            |19.2 |6   |167.6 |123 |3.92 |3.44  |18.3  |1  |0  |4    |4    |
#> |Merc 280C           |17.8 |6   |167.6 |123 |3.92 |3.44  |18.9  |1  |0  |4    |4    |
#> |Merc 450SE          |16.4 |8   |275.8 |180 |3.07 |4.07  |17.4  |0  |0  |3    |3    |
#> |Merc 450SL          |17.3 |8   |275.8 |180 |3.07 |3.73  |17.6  |0  |0  |3    |3    |
#> |Merc 450SLC         |15.2 |8   |275.8 |180 |3.07 |3.78  |18    |0  |0  |3    |3    |
#> |Cadillac Fleetwood  |10.4 |8   |472   |205 |2.93 |5.25  |17.98 |0  |0  |3    |4    |
#> |Lincoln Continental |10.4 |8   |460   |215 |3    |5.424 |17.82 |0  |0  |3    |4    |
#> |Chrysler Imperial   |14.7 |8   |440   |230 |3.23 |5.345 |17.42 |0  |0  |3    |4    |
#> |Fiat 128            |32.4 |4   |78.7  |66  |4.08 |2.2   |19.47 |1  |1  |4    |1    |
#> |Honda Civic         |30.4 |4   |75.7  |52  |4.93 |1.615 |18.52 |1  |1  |4    |2    |
#> |Toyota Corolla      |33.9 |4   |71.1  |65  |4.22 |1.835 |19.9  |1  |1  |4    |1    |
#> |Toyota Corona       |21.5 |4   |120.1 |97  |3.7  |2.465 |20.01 |1  |0  |3    |1    |
#> |Dodge Challenger    |15.5 |8   |318   |150 |2.76 |3.52  |16.87 |0  |0  |3    |2    |
#> |AMC Javelin         |15.2 |8   |304   |150 |3.15 |3.435 |17.3  |0  |0  |3    |2    |
#> |Camaro Z28          |13.3 |8   |350   |245 |3.73 |3.84  |15.41 |0  |0  |3    |4    |
#> |Pontiac Firebird    |19.2 |8   |400   |175 |3.08 |3.845 |17.05 |0  |0  |3    |2    |
#> |Fiat X1-9           |27.3 |4   |79    |66  |4.08 |1.935 |18.9  |1  |1  |4    |1    |
#> |Porsche 914-2       |26   |4   |120.3 |91  |4.43 |2.14  |16.7  |0  |1  |5    |2    |
#> |Lotus Europa        |30.4 |4   |95.1  |113 |3.77 |1.513 |16.9  |1  |1  |5    |2    |
#> |Ford Pantera L      |15.8 |8   |351   |264 |4.22 |3.17  |14.5  |0  |1  |5    |4    |
#> |Ferrari Dino        |19.7 |6   |145   |175 |3.62 |2.77  |15.5  |0  |1  |5    |6    |
#> |Maserati Bora       |15   |8   |301   |335 |3.54 |3.57  |14.6  |0  |1  |5    |8    |
#> |Volvo 142E          |21.4 |4   |121   |109 |4.11 |2.78  |18.6  |1  |1  |4    |2    |

# by make
make <- sub("^(\\w+)\\s?(.*)$", "\\1", rownames(mtcars))
make <- c(table(make))

# A LaTeX table with a vertical bar between each column
qable(mtcars[sort(rownames(mtcars)), ], rgroup = make)
#> 
#> \begin{tabular}{l|l|l|l|l|l|l|l|l|l|l|l}
#> \hline
#>  & mpg & cyl & disp & hp & drat & wt & qsec & vs & am & gear & carb\\
#> \hline
#> \bf{AMC} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ AMC Javelin & 15.2 & 8 & 304 & 150 & 3.15 & 3.435 & 17.3 & 0 & 0 & 3 & 2\\
#> \hline
#> \bf{Cadillac} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Cadillac Fleetwood & 10.4 & 8 & 472 & 205 & 2.93 & 5.25 & 17.98 & 0 & 0 & 3 & 4\\
#> \hline
#> \bf{Camaro} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Camaro Z28 & 13.3 & 8 & 350 & 245 & 3.73 & 3.84 & 15.41 & 0 & 0 & 3 & 4\\
#> \hline
#> \bf{Chrysler} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Chrysler Imperial & 14.7 & 8 & 440 & 230 & 3.23 & 5.345 & 17.42 & 0 & 0 & 3 & 4\\
#> \hline
#> \bf{Datsun} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Datsun 710 & 22.8 & 4 & 108 & 93 & 3.85 & 2.32 & 18.61 & 1 & 1 & 4 & 1\\
#> \hline
#> \bf{Dodge} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Dodge Challenger & 15.5 & 8 & 318 & 150 & 2.76 & 3.52 & 16.87 & 0 & 0 & 3 & 2\\
#> \hline
#> \bf{Duster} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Duster 360 & 14.3 & 8 & 360 & 245 & 3.21 & 3.57 & 15.84 & 0 & 0 & 3 & 4\\
#> \hline
#> \bf{Ferrari} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Ferrari Dino & 19.7 & 6 & 145 & 175 & 3.62 & 2.77 & 15.5 & 0 & 1 & 5 & 6\\
#> \hline
#> \bf{Fiat} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Fiat 128 & 32.4 & 4 & 78.7 & 66 & 4.08 & 2.2 & 19.47 & 1 & 1 & 4 & 1\\
#> \hline
#> ~~ Fiat X1-9 & 27.3 & 4 & 79 & 66 & 4.08 & 1.935 & 18.9 & 1 & 1 & 4 & 1\\
#> \hline
#> \bf{Ford} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Ford Pantera L & 15.8 & 8 & 351 & 264 & 4.22 & 3.17 & 14.5 & 0 & 1 & 5 & 4\\
#> \hline
#> \bf{Honda} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Honda Civic & 30.4 & 4 & 75.7 & 52 & 4.93 & 1.615 & 18.52 & 1 & 1 & 4 & 2\\
#> \hline
#> \bf{Hornet} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Hornet 4 Drive & 21.4 & 6 & 258 & 110 & 3.08 & 3.215 & 19.44 & 1 & 0 & 3 & 1\\
#> \hline
#> ~~ Hornet Sportabout & 18.7 & 8 & 360 & 175 & 3.15 & 3.44 & 17.02 & 0 & 0 & 3 & 2\\
#> \hline
#> \bf{Lincoln} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Lincoln Continental & 10.4 & 8 & 460 & 215 & 3 & 5.424 & 17.82 & 0 & 0 & 3 & 4\\
#> \hline
#> \bf{Lotus} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Lotus Europa & 30.4 & 4 & 95.1 & 113 & 3.77 & 1.513 & 16.9 & 1 & 1 & 5 & 2\\
#> \hline
#> \bf{Maserati} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Maserati Bora & 15 & 8 & 301 & 335 & 3.54 & 3.57 & 14.6 & 0 & 1 & 5 & 8\\
#> \hline
#> \bf{Mazda} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Mazda RX4 & 21 & 6 & 160 & 110 & 3.9 & 2.62 & 16.46 & 0 & 1 & 4 & 4\\
#> \hline
#> ~~ Mazda RX4 Wag & 21 & 6 & 160 & 110 & 3.9 & 2.875 & 17.02 & 0 & 1 & 4 & 4\\
#> \hline
#> \bf{Merc} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Merc 230 & 22.8 & 4 & 140.8 & 95 & 3.92 & 3.15 & 22.9 & 1 & 0 & 4 & 2\\
#> \hline
#> ~~ Merc 240D & 24.4 & 4 & 146.7 & 62 & 3.69 & 3.19 & 20 & 1 & 0 & 4 & 2\\
#> \hline
#> ~~ Merc 280 & 19.2 & 6 & 167.6 & 123 & 3.92 & 3.44 & 18.3 & 1 & 0 & 4 & 4\\
#> \hline
#> ~~ Merc 280C & 17.8 & 6 & 167.6 & 123 & 3.92 & 3.44 & 18.9 & 1 & 0 & 4 & 4\\
#> \hline
#> ~~ Merc 450SE & 16.4 & 8 & 275.8 & 180 & 3.07 & 4.07 & 17.4 & 0 & 0 & 3 & 3\\
#> \hline
#> ~~ Merc 450SL & 17.3 & 8 & 275.8 & 180 & 3.07 & 3.73 & 17.6 & 0 & 0 & 3 & 3\\
#> \hline
#> ~~ Merc 450SLC & 15.2 & 8 & 275.8 & 180 & 3.07 & 3.78 & 18 & 0 & 0 & 3 & 3\\
#> \hline
#> \bf{Pontiac} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Pontiac Firebird & 19.2 & 8 & 400 & 175 & 3.08 & 3.845 & 17.05 & 0 & 0 & 3 & 2\\
#> \hline
#> \bf{Porsche} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Porsche 914-2 & 26 & 4 & 120.3 & 91 & 4.43 & 2.14 & 16.7 & 0 & 1 & 5 & 2\\
#> \hline
#> \bf{Toyota} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Toyota Corolla & 33.9 & 4 & 71.1 & 65 & 4.22 & 1.835 & 19.9 & 1 & 1 & 4 & 1\\
#> \hline
#> ~~ Toyota Corona & 21.5 & 4 & 120.1 & 97 & 3.7 & 2.465 & 20.01 & 1 & 0 & 3 & 1\\
#> \hline
#> \bf{Valiant} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Valiant & 18.1 & 6 & 225 & 105 & 2.76 & 3.46 & 20.22 & 1 & 0 & 3 & 1\\
#> \hline
#> \bf{Volvo} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Volvo 142E & 21.4 & 4 & 121 & 109 & 4.11 & 2.78 & 18.6 & 1 & 1 & 4 & 2\\
#> \hline
#> \end{tabular}

# A LaTeX table with no vertical bars between columns
qable(mtcars[sort(rownames(mtcars)), ], rgroup = make, kable_args = list(vline = ""))
#> 
#> \begin{tabular}{llllllllllll}
#> \hline
#>  & mpg & cyl & disp & hp & drat & wt & qsec & vs & am & gear & carb\\
#> \hline
#> \bf{AMC} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ AMC Javelin & 15.2 & 8 & 304 & 150 & 3.15 & 3.435 & 17.3 & 0 & 0 & 3 & 2\\
#> \hline
#> \bf{Cadillac} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Cadillac Fleetwood & 10.4 & 8 & 472 & 205 & 2.93 & 5.25 & 17.98 & 0 & 0 & 3 & 4\\
#> \hline
#> \bf{Camaro} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Camaro Z28 & 13.3 & 8 & 350 & 245 & 3.73 & 3.84 & 15.41 & 0 & 0 & 3 & 4\\
#> \hline
#> \bf{Chrysler} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Chrysler Imperial & 14.7 & 8 & 440 & 230 & 3.23 & 5.345 & 17.42 & 0 & 0 & 3 & 4\\
#> \hline
#> \bf{Datsun} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Datsun 710 & 22.8 & 4 & 108 & 93 & 3.85 & 2.32 & 18.61 & 1 & 1 & 4 & 1\\
#> \hline
#> \bf{Dodge} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Dodge Challenger & 15.5 & 8 & 318 & 150 & 2.76 & 3.52 & 16.87 & 0 & 0 & 3 & 2\\
#> \hline
#> \bf{Duster} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Duster 360 & 14.3 & 8 & 360 & 245 & 3.21 & 3.57 & 15.84 & 0 & 0 & 3 & 4\\
#> \hline
#> \bf{Ferrari} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Ferrari Dino & 19.7 & 6 & 145 & 175 & 3.62 & 2.77 & 15.5 & 0 & 1 & 5 & 6\\
#> \hline
#> \bf{Fiat} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Fiat 128 & 32.4 & 4 & 78.7 & 66 & 4.08 & 2.2 & 19.47 & 1 & 1 & 4 & 1\\
#> \hline
#> ~~ Fiat X1-9 & 27.3 & 4 & 79 & 66 & 4.08 & 1.935 & 18.9 & 1 & 1 & 4 & 1\\
#> \hline
#> \bf{Ford} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Ford Pantera L & 15.8 & 8 & 351 & 264 & 4.22 & 3.17 & 14.5 & 0 & 1 & 5 & 4\\
#> \hline
#> \bf{Honda} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Honda Civic & 30.4 & 4 & 75.7 & 52 & 4.93 & 1.615 & 18.52 & 1 & 1 & 4 & 2\\
#> \hline
#> \bf{Hornet} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Hornet 4 Drive & 21.4 & 6 & 258 & 110 & 3.08 & 3.215 & 19.44 & 1 & 0 & 3 & 1\\
#> \hline
#> ~~ Hornet Sportabout & 18.7 & 8 & 360 & 175 & 3.15 & 3.44 & 17.02 & 0 & 0 & 3 & 2\\
#> \hline
#> \bf{Lincoln} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Lincoln Continental & 10.4 & 8 & 460 & 215 & 3 & 5.424 & 17.82 & 0 & 0 & 3 & 4\\
#> \hline
#> \bf{Lotus} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Lotus Europa & 30.4 & 4 & 95.1 & 113 & 3.77 & 1.513 & 16.9 & 1 & 1 & 5 & 2\\
#> \hline
#> \bf{Maserati} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Maserati Bora & 15 & 8 & 301 & 335 & 3.54 & 3.57 & 14.6 & 0 & 1 & 5 & 8\\
#> \hline
#> \bf{Mazda} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Mazda RX4 & 21 & 6 & 160 & 110 & 3.9 & 2.62 & 16.46 & 0 & 1 & 4 & 4\\
#> \hline
#> ~~ Mazda RX4 Wag & 21 & 6 & 160 & 110 & 3.9 & 2.875 & 17.02 & 0 & 1 & 4 & 4\\
#> \hline
#> \bf{Merc} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Merc 230 & 22.8 & 4 & 140.8 & 95 & 3.92 & 3.15 & 22.9 & 1 & 0 & 4 & 2\\
#> \hline
#> ~~ Merc 240D & 24.4 & 4 & 146.7 & 62 & 3.69 & 3.19 & 20 & 1 & 0 & 4 & 2\\
#> \hline
#> ~~ Merc 280 & 19.2 & 6 & 167.6 & 123 & 3.92 & 3.44 & 18.3 & 1 & 0 & 4 & 4\\
#> \hline
#> ~~ Merc 280C & 17.8 & 6 & 167.6 & 123 & 3.92 & 3.44 & 18.9 & 1 & 0 & 4 & 4\\
#> \hline
#> ~~ Merc 450SE & 16.4 & 8 & 275.8 & 180 & 3.07 & 4.07 & 17.4 & 0 & 0 & 3 & 3\\
#> \hline
#> ~~ Merc 450SL & 17.3 & 8 & 275.8 & 180 & 3.07 & 3.73 & 17.6 & 0 & 0 & 3 & 3\\
#> \hline
#> ~~ Merc 450SLC & 15.2 & 8 & 275.8 & 180 & 3.07 & 3.78 & 18 & 0 & 0 & 3 & 3\\
#> \hline
#> \bf{Pontiac} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Pontiac Firebird & 19.2 & 8 & 400 & 175 & 3.08 & 3.845 & 17.05 & 0 & 0 & 3 & 2\\
#> \hline
#> \bf{Porsche} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Porsche 914-2 & 26 & 4 & 120.3 & 91 & 4.43 & 2.14 & 16.7 & 0 & 1 & 5 & 2\\
#> \hline
#> \bf{Toyota} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Toyota Corolla & 33.9 & 4 & 71.1 & 65 & 4.22 & 1.835 & 19.9 & 1 & 1 & 4 & 1\\
#> \hline
#> ~~ Toyota Corona & 21.5 & 4 & 120.1 & 97 & 3.7 & 2.465 & 20.01 & 1 & 0 & 3 & 1\\
#> \hline
#> \bf{Valiant} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Valiant & 18.1 & 6 & 225 & 105 & 2.76 & 3.46 & 20.22 & 1 & 0 & 3 & 1\\
#> \hline
#> \bf{Volvo} & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~ & ~\\
#> \hline
#> ~~ Volvo 142E & 21.4 & 4 & 121 & 109 & 4.11 & 2.78 & 18.6 & 1 & 1 & 4 & 2\\
#> \hline
#> \end{tabular}

# a markdown table
qable(mtcars[sort(rownames(mtcars)), ], rgroup = make, markup = "markdown")
#> 
#> 
#> |                                 |mpg          |cyl          |disp         |hp           |drat         |wt           |qsec         |vs           |am           |gear         |carb         |
#> |:--------------------------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|
#> |**AMC**                          |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; AMC Javelin         |15.2         |8            |304          |150          |3.15         |3.435        |17.3         |0            |0            |3            |2            |
#> |**Cadillac**                     |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Cadillac Fleetwood  |10.4         |8            |472          |205          |2.93         |5.25         |17.98        |0            |0            |3            |4            |
#> |**Camaro**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Camaro Z28          |13.3         |8            |350          |245          |3.73         |3.84         |15.41        |0            |0            |3            |4            |
#> |**Chrysler**                     |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Chrysler Imperial   |14.7         |8            |440          |230          |3.23         |5.345        |17.42        |0            |0            |3            |4            |
#> |**Datsun**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Datsun 710          |22.8         |4            |108          |93           |3.85         |2.32         |18.61        |1            |1            |4            |1            |
#> |**Dodge**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Dodge Challenger    |15.5         |8            |318          |150          |2.76         |3.52         |16.87        |0            |0            |3            |2            |
#> |**Duster**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Duster 360          |14.3         |8            |360          |245          |3.21         |3.57         |15.84        |0            |0            |3            |4            |
#> |**Ferrari**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Ferrari Dino        |19.7         |6            |145          |175          |3.62         |2.77         |15.5         |0            |1            |5            |6            |
#> |**Fiat**                         |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Fiat 128            |32.4         |4            |78.7         |66           |4.08         |2.2          |19.47        |1            |1            |4            |1            |
#> |&nbsp;&nbsp; Fiat X1-9           |27.3         |4            |79           |66           |4.08         |1.935        |18.9         |1            |1            |4            |1            |
#> |**Ford**                         |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Ford Pantera L      |15.8         |8            |351          |264          |4.22         |3.17         |14.5         |0            |1            |5            |4            |
#> |**Honda**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Honda Civic         |30.4         |4            |75.7         |52           |4.93         |1.615        |18.52        |1            |1            |4            |2            |
#> |**Hornet**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Hornet 4 Drive      |21.4         |6            |258          |110          |3.08         |3.215        |19.44        |1            |0            |3            |1            |
#> |&nbsp;&nbsp; Hornet Sportabout   |18.7         |8            |360          |175          |3.15         |3.44         |17.02        |0            |0            |3            |2            |
#> |**Lincoln**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Lincoln Continental |10.4         |8            |460          |215          |3            |5.424        |17.82        |0            |0            |3            |4            |
#> |**Lotus**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Lotus Europa        |30.4         |4            |95.1         |113          |3.77         |1.513        |16.9         |1            |1            |5            |2            |
#> |**Maserati**                     |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Maserati Bora       |15           |8            |301          |335          |3.54         |3.57         |14.6         |0            |1            |5            |8            |
#> |**Mazda**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Mazda RX4           |21           |6            |160          |110          |3.9          |2.62         |16.46        |0            |1            |4            |4            |
#> |&nbsp;&nbsp; Mazda RX4 Wag       |21           |6            |160          |110          |3.9          |2.875        |17.02        |0            |1            |4            |4            |
#> |**Merc**                         |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Merc 230            |22.8         |4            |140.8        |95           |3.92         |3.15         |22.9         |1            |0            |4            |2            |
#> |&nbsp;&nbsp; Merc 240D           |24.4         |4            |146.7        |62           |3.69         |3.19         |20           |1            |0            |4            |2            |
#> |&nbsp;&nbsp; Merc 280            |19.2         |6            |167.6        |123          |3.92         |3.44         |18.3         |1            |0            |4            |4            |
#> |&nbsp;&nbsp; Merc 280C           |17.8         |6            |167.6        |123          |3.92         |3.44         |18.9         |1            |0            |4            |4            |
#> |&nbsp;&nbsp; Merc 450SE          |16.4         |8            |275.8        |180          |3.07         |4.07         |17.4         |0            |0            |3            |3            |
#> |&nbsp;&nbsp; Merc 450SL          |17.3         |8            |275.8        |180          |3.07         |3.73         |17.6         |0            |0            |3            |3            |
#> |&nbsp;&nbsp; Merc 450SLC         |15.2         |8            |275.8        |180          |3.07         |3.78         |18           |0            |0            |3            |3            |
#> |**Pontiac**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Pontiac Firebird    |19.2         |8            |400          |175          |3.08         |3.845        |17.05        |0            |0            |3            |2            |
#> |**Porsche**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Porsche 914-2       |26           |4            |120.3        |91           |4.43         |2.14         |16.7         |0            |1            |5            |2            |
#> |**Toyota**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Toyota Corolla      |33.9         |4            |71.1         |65           |4.22         |1.835        |19.9         |1            |1            |4            |1            |
#> |&nbsp;&nbsp; Toyota Corona       |21.5         |4            |120.1        |97           |3.7          |2.465        |20.01        |1            |0            |3            |1            |
#> |**Valiant**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Valiant             |18.1         |6            |225          |105          |2.76         |3.46         |20.22        |1            |0            |3            |1            |
#> |**Volvo**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Volvo 142E          |21.4         |4            |121          |109          |4.11         |2.78         |18.6         |1            |1            |4            |2            |

# define your own column names
qable(mtcars[sort(rownames(mtcars)), ],
      rgroup = make,
      cnames = toupper(colnames(mtcars)),
      markup = "markdown")
#> 
#> 
#> |                                 |MPG          |CYL          |DISP         |HP           |DRAT         |WT           |QSEC         |VS           |AM           |GEAR         |CARB         |
#> |:--------------------------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|
#> |**AMC**                          |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; AMC Javelin         |15.2         |8            |304          |150          |3.15         |3.435        |17.3         |0            |0            |3            |2            |
#> |**Cadillac**                     |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Cadillac Fleetwood  |10.4         |8            |472          |205          |2.93         |5.25         |17.98        |0            |0            |3            |4            |
#> |**Camaro**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Camaro Z28          |13.3         |8            |350          |245          |3.73         |3.84         |15.41        |0            |0            |3            |4            |
#> |**Chrysler**                     |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Chrysler Imperial   |14.7         |8            |440          |230          |3.23         |5.345        |17.42        |0            |0            |3            |4            |
#> |**Datsun**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Datsun 710          |22.8         |4            |108          |93           |3.85         |2.32         |18.61        |1            |1            |4            |1            |
#> |**Dodge**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Dodge Challenger    |15.5         |8            |318          |150          |2.76         |3.52         |16.87        |0            |0            |3            |2            |
#> |**Duster**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Duster 360          |14.3         |8            |360          |245          |3.21         |3.57         |15.84        |0            |0            |3            |4            |
#> |**Ferrari**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Ferrari Dino        |19.7         |6            |145          |175          |3.62         |2.77         |15.5         |0            |1            |5            |6            |
#> |**Fiat**                         |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Fiat 128            |32.4         |4            |78.7         |66           |4.08         |2.2          |19.47        |1            |1            |4            |1            |
#> |&nbsp;&nbsp; Fiat X1-9           |27.3         |4            |79           |66           |4.08         |1.935        |18.9         |1            |1            |4            |1            |
#> |**Ford**                         |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Ford Pantera L      |15.8         |8            |351          |264          |4.22         |3.17         |14.5         |0            |1            |5            |4            |
#> |**Honda**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Honda Civic         |30.4         |4            |75.7         |52           |4.93         |1.615        |18.52        |1            |1            |4            |2            |
#> |**Hornet**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Hornet 4 Drive      |21.4         |6            |258          |110          |3.08         |3.215        |19.44        |1            |0            |3            |1            |
#> |&nbsp;&nbsp; Hornet Sportabout   |18.7         |8            |360          |175          |3.15         |3.44         |17.02        |0            |0            |3            |2            |
#> |**Lincoln**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Lincoln Continental |10.4         |8            |460          |215          |3            |5.424        |17.82        |0            |0            |3            |4            |
#> |**Lotus**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Lotus Europa        |30.4         |4            |95.1         |113          |3.77         |1.513        |16.9         |1            |1            |5            |2            |
#> |**Maserati**                     |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Maserati Bora       |15           |8            |301          |335          |3.54         |3.57         |14.6         |0            |1            |5            |8            |
#> |**Mazda**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Mazda RX4           |21           |6            |160          |110          |3.9          |2.62         |16.46        |0            |1            |4            |4            |
#> |&nbsp;&nbsp; Mazda RX4 Wag       |21           |6            |160          |110          |3.9          |2.875        |17.02        |0            |1            |4            |4            |
#> |**Merc**                         |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Merc 230            |22.8         |4            |140.8        |95           |3.92         |3.15         |22.9         |1            |0            |4            |2            |
#> |&nbsp;&nbsp; Merc 240D           |24.4         |4            |146.7        |62           |3.69         |3.19         |20           |1            |0            |4            |2            |
#> |&nbsp;&nbsp; Merc 280            |19.2         |6            |167.6        |123          |3.92         |3.44         |18.3         |1            |0            |4            |4            |
#> |&nbsp;&nbsp; Merc 280C           |17.8         |6            |167.6        |123          |3.92         |3.44         |18.9         |1            |0            |4            |4            |
#> |&nbsp;&nbsp; Merc 450SE          |16.4         |8            |275.8        |180          |3.07         |4.07         |17.4         |0            |0            |3            |3            |
#> |&nbsp;&nbsp; Merc 450SL          |17.3         |8            |275.8        |180          |3.07         |3.73         |17.6         |0            |0            |3            |3            |
#> |&nbsp;&nbsp; Merc 450SLC         |15.2         |8            |275.8        |180          |3.07         |3.78         |18           |0            |0            |3            |3            |
#> |**Pontiac**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Pontiac Firebird    |19.2         |8            |400          |175          |3.08         |3.845        |17.05        |0            |0            |3            |2            |
#> |**Porsche**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Porsche 914-2       |26           |4            |120.3        |91           |4.43         |2.14         |16.7         |0            |1            |5            |2            |
#> |**Toyota**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Toyota Corolla      |33.9         |4            |71.1         |65           |4.22         |1.835        |19.9         |1            |1            |4            |1            |
#> |&nbsp;&nbsp; Toyota Corona       |21.5         |4            |120.1        |97           |3.7          |2.465        |20.01        |1            |0            |3            |1            |
#> |**Valiant**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Valiant             |18.1         |6            |225          |105          |2.76         |3.46         |20.22        |1            |0            |3            |1            |
#> |**Volvo**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Volvo 142E          |21.4         |4            |121          |109          |4.11         |2.78         |18.6         |1            |1            |4            |2            |

# define your own column names and add a title
qable(mtcars[sort(rownames(mtcars)), ],
      rtitle = "Make & Model",
      rgroup = make,
      cnames = toupper(colnames(mtcars)),
      markup = "markdown")
#> 
#> 
#> |Make & Model                     |MPG          |CYL          |DISP         |HP           |DRAT         |WT           |QSEC         |VS           |AM           |GEAR         |CARB         |
#> |:--------------------------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|:------------|
#> |**AMC**                          |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; AMC Javelin         |15.2         |8            |304          |150          |3.15         |3.435        |17.3         |0            |0            |3            |2            |
#> |**Cadillac**                     |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Cadillac Fleetwood  |10.4         |8            |472          |205          |2.93         |5.25         |17.98        |0            |0            |3            |4            |
#> |**Camaro**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Camaro Z28          |13.3         |8            |350          |245          |3.73         |3.84         |15.41        |0            |0            |3            |4            |
#> |**Chrysler**                     |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Chrysler Imperial   |14.7         |8            |440          |230          |3.23         |5.345        |17.42        |0            |0            |3            |4            |
#> |**Datsun**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Datsun 710          |22.8         |4            |108          |93           |3.85         |2.32         |18.61        |1            |1            |4            |1            |
#> |**Dodge**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Dodge Challenger    |15.5         |8            |318          |150          |2.76         |3.52         |16.87        |0            |0            |3            |2            |
#> |**Duster**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Duster 360          |14.3         |8            |360          |245          |3.21         |3.57         |15.84        |0            |0            |3            |4            |
#> |**Ferrari**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Ferrari Dino        |19.7         |6            |145          |175          |3.62         |2.77         |15.5         |0            |1            |5            |6            |
#> |**Fiat**                         |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Fiat 128            |32.4         |4            |78.7         |66           |4.08         |2.2          |19.47        |1            |1            |4            |1            |
#> |&nbsp;&nbsp; Fiat X1-9           |27.3         |4            |79           |66           |4.08         |1.935        |18.9         |1            |1            |4            |1            |
#> |**Ford**                         |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Ford Pantera L      |15.8         |8            |351          |264          |4.22         |3.17         |14.5         |0            |1            |5            |4            |
#> |**Honda**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Honda Civic         |30.4         |4            |75.7         |52           |4.93         |1.615        |18.52        |1            |1            |4            |2            |
#> |**Hornet**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Hornet 4 Drive      |21.4         |6            |258          |110          |3.08         |3.215        |19.44        |1            |0            |3            |1            |
#> |&nbsp;&nbsp; Hornet Sportabout   |18.7         |8            |360          |175          |3.15         |3.44         |17.02        |0            |0            |3            |2            |
#> |**Lincoln**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Lincoln Continental |10.4         |8            |460          |215          |3            |5.424        |17.82        |0            |0            |3            |4            |
#> |**Lotus**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Lotus Europa        |30.4         |4            |95.1         |113          |3.77         |1.513        |16.9         |1            |1            |5            |2            |
#> |**Maserati**                     |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Maserati Bora       |15           |8            |301          |335          |3.54         |3.57         |14.6         |0            |1            |5            |8            |
#> |**Mazda**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Mazda RX4           |21           |6            |160          |110          |3.9          |2.62         |16.46        |0            |1            |4            |4            |
#> |&nbsp;&nbsp; Mazda RX4 Wag       |21           |6            |160          |110          |3.9          |2.875        |17.02        |0            |1            |4            |4            |
#> |**Merc**                         |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Merc 230            |22.8         |4            |140.8        |95           |3.92         |3.15         |22.9         |1            |0            |4            |2            |
#> |&nbsp;&nbsp; Merc 240D           |24.4         |4            |146.7        |62           |3.69         |3.19         |20           |1            |0            |4            |2            |
#> |&nbsp;&nbsp; Merc 280            |19.2         |6            |167.6        |123          |3.92         |3.44         |18.3         |1            |0            |4            |4            |
#> |&nbsp;&nbsp; Merc 280C           |17.8         |6            |167.6        |123          |3.92         |3.44         |18.9         |1            |0            |4            |4            |
#> |&nbsp;&nbsp; Merc 450SE          |16.4         |8            |275.8        |180          |3.07         |4.07         |17.4         |0            |0            |3            |3            |
#> |&nbsp;&nbsp; Merc 450SL          |17.3         |8            |275.8        |180          |3.07         |3.73         |17.6         |0            |0            |3            |3            |
#> |&nbsp;&nbsp; Merc 450SLC         |15.2         |8            |275.8        |180          |3.07         |3.78         |18           |0            |0            |3            |3            |
#> |**Pontiac**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Pontiac Firebird    |19.2         |8            |400          |175          |3.08         |3.845        |17.05        |0            |0            |3            |2            |
#> |**Porsche**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Porsche 914-2       |26           |4            |120.3        |91           |4.43         |2.14         |16.7         |0            |1            |5            |2            |
#> |**Toyota**                       |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Toyota Corolla      |33.9         |4            |71.1         |65           |4.22         |1.835        |19.9         |1            |1            |4            |1            |
#> |&nbsp;&nbsp; Toyota Corona       |21.5         |4            |120.1        |97           |3.7          |2.465        |20.01        |1            |0            |3            |1            |
#> |**Valiant**                      |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Valiant             |18.1         |6            |225          |105          |2.76         |3.46         |20.22        |1            |0            |3            |1            |
#> |**Volvo**                        |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |&nbsp;&nbsp; |
#> |&nbsp;&nbsp; Volvo 142E          |21.4         |4            |121          |109          |4.11         |2.78         |18.6         |1            |1            |4            |2            |